Use rhel9 base image in Dockerfile
Recently, we added OKD configs for openshift components. The config overrides the base image with a centos image, which doesn't work because the base image for the Dockerfile is an nginx image. To be consistent with other repos and the ART dockerfile, change the nginx base image to a rhel base image, so that prow cando the substitution. This also prevents any CVE vulnerabilities we would have when using the nginx image.